feat(nested-battle-menu): Adds a nested battle menu
I want to be able to pick skills in a menu! Let's create a nested battle menu
that's capable of finding the information we want about a unit (once we have that information).

In order to get things to overlap in the way I want, I had to switch to using CSS Grids.
I ended up rewriting the majority of the Battle Menu logic while I was here :shrug: